GO-5847: Extract object open machinery to component
