+++ core/block/editor/smartblock/links.go
@@ -3,9 +3,7 @@ package smartblock
 import (
 	"github.com/anyproto/anytype-heart/core/block/editor/state"
 	"github.com/anyproto/anytype-heart/core/block/object/objectlink"
-	"github.com/anyproto/anytype-heart/core/block/simple"
 	"github.com/anyproto/anytype-heart/core/domain"
-	"github.com/anyproto/anytype-heart/pb"
 	"github.com/anyproto/anytype-heart/pkg/lib/bundle"
 	"github.com/anyproto/anytype-heart/pkg/lib/pb/model"
 	"github.com/anyproto/anytype-heart/util/internalflag"
@@ -23,9 +21,10 @@ func (sb *smartBlock) updateBackLinks(s *state.State) {
 
 func (sb *smartBlock) injectLinksDetails(s *state.State) {
 	links := objectlink.DependentObjectIDs(s, sb.Space(), objectlink.Flags{
-		Blocks:                   true,
-		Details:                  true,
-		Relations:                sb.includeRelationObjectsAsDependents,
+		Blocks:  true,
+		Details: true,
+		// TODO Why here?
+		// Relations:                sb.includeRelationObjectsAsDependents,
 		Types:                    false,
 		Collection:               !internalflag.NewFromState(s).Has(model.InternalFlag_collectionDontIndexLinks),
 		DataviewBlockOnlyTarget:  true,
@@ -52,17 +51,4 @@ func (sb *smartBlock) injectMentions(s *state.State) {
 	})
 	mentions = slice.RemoveMut(mentions, sb.Id())
 	s.SetDetailAndBundledRelation(bundle.RelationKeyMentions, domain.StringList(mentions))
-}
-
-func isBacklinksChanged(msgs []simple.EventMessage) bool {
-	for _, msg := range msgs {
-		if amend, ok := msg.Msg.Value.(*pb.EventMessageValueOfObjectDetailsAmend); ok {
-			for _, detail := range amend.ObjectDetailsAmend.Details {
-				if detail.Key == bundle.RelationKeyBacklinks.String() {
-					return true
-				}
-			}
-		}
-	}
-	return false
 }
